requirements.txt
pymongo--> MongoClient, ServerApi
os
flask--> Blueprint, request, jsonify, session, redirect, url_for, render_template
werkzeug.security --> enerate_password_hash, check_password_hash